Components moved to corresponding files. Added PureComponent.
