+++ src/counter.jsx
@@ -1,0 +1,32 @@
+import React from "react";
+
+class Counter extends React.Component {
+  constructor() {
+    super();
+    this.state = {
+      counter: 0,
+    };
+    this.increment = this.increment.bind(this);
+    this.decrement = this.decrement.bind(this);
+  }
+
+  increment() {
+    this.setState((state) => ({ counter: state.counter + 1 }));
+  }
+
+  decrement() {
+    this.setState((state) => ({ counter: state.counter - 1 }));
+  }
+
+  render() {
+    return (
+      <div>
+        <h1>{this.state.counter}</h1>
+        <button onClick={this.increment}>+</button>
+        <button onClick={this.decrement}>-</button>
+      </div>
+    );
+  }
+}
+
+export default Counter;
